Successfully sending login/auth between client/server
Made many errors, none too egregious. This time the bug was
that the pipeline order client side was incorrect; the decoder
was put after the AuthHandler so that every time the client
Auth Handler was trying to read the Message data an Index
Out of Bounds exception was thrown! Simple mistake, lots of
heartache.