Work on browser client HTML Post; I hate Angular
I spent about three hours trying to learn Angular.js to make
a simple implementation of an HTML POST request for the login
form that I have. It was way too confusing for me and I think
right now it's more important to have a working product.

I went for the simple route with a basic JS listener for
the form on submit. Need to set up server HTTP handler so
that it can handle the websocket post request?