Added Protocol Handler
To send login credentials over the wire using the browser client,
I need to use HTTP post. Browser clients cannot send POST over the
websockets protocol so the server needs a way of identifying if it's
receiving a FlatBuffers packet or a POST request for login BEFORE
adding in the WS pipeline in the channel. Therefore I've added a
Protocol Handler that should identify the type of request the server
is receiving and allocate the correct handlers afterwards to deal with
the protocol that the client is forwarding the credentials with.