Fixed browser WS close delay bug
Hacked together a solution by changing the UI to display that
the connection was closed before performing cleanup for closing
the websocket and removing the interval functions (pinger,
statusChecker).

This is because there is a delay between the time the WebSocket
is closed and the time the WebSocket onClose function is actually
executed for whatever reason. On the bright side, the server knows
that the connection is closed and the status of the ws client side
is closed as soon as ws.close() is called.

Also put cleanup code into a function called closeWs().

Documentation to follow.