Fixed server response bug w/ content length header
Needed to add content length header to the HTTP response.

The client was receiving the messages but with no content
length header specified, kept waiting for the full message to be
received. This explains why it would show the messages on
server connection close - that would be confirmation that
there was no more message to be received.