Websockets Protocol Handled; Session DB
Added Session DB for authenticated users to be checked
before completing WS handshake

Users should be authenticated before attempting handshake.
Upon attempting handshake, the server will check to make sure
that the user's session is valid and complete the handshake if so.

Also altered the protocol handler to filter GET requests as WS,
as the only other HTTP request we should be dealing with is POST
for login and OPTIONS in the case that the HTTP Post request
needs CORS.