Changing websockets login to include ticket param
Websocket login URI changed so that we add a ticket param
to the end of the websockets handshake request, i.e.

var ws = new WebSockets("ws://someUrl/whatever?ticket=xxxxx");

The ticket will be verified the server and it will be used to determine
whether or not the handshake will complete successfully.

Tickets are obtained from login requests which should be attempted before
WS handshake.