Websockets service, ui-router, ticket
Added WebSockets service so that all controllers have access to a
singleton websocket instance for the web application.

Also implemented ui-router change so that the view changes when the
user logs in to the server. Currently the only thing that happens
is the view removes login.html and replaces it with chat.html, also
updating the current url. chat.html is not functional currently
so this is purely cosmetic.

Implemented ticket retrieval from server and immediate login with
ticket using WebSockets after the ticket is received. This was
achieved appending the ticket to the WebSockets handshake query string.
More details are available in websocketsService.js

Will change name of service and do more organizing + documentation later.