Chat for web app sending and receiving server messages
Implemented ChatSendController for chat.html, responsible for sending
chat messages. Current implementation also receives the messages but
that function will be relegated to a different controller or part of
chat.html which will display the messages.

Changed ctor of ChatServerHandler so that it adds the current channel
to the group of all channels when it is called so that users are registered
in the global channelgroup.

Need to test if it works with Java client next.