Refactored Chat Controllers. Server and client read/write lobbies
Chat controllers refactored into separate files, index.html updated
dependencies.

Redid ChatReceiveController and chat.html so that the array holding
server messages which is showed in the view consists of strings
as opposed to message objects. This is so that the client can
receive many types of messages and display them in the view.

Specifically it was to facilitate the implementation of the
receiving of Lobby lists by the client. Web client can now
receive a list of the lobbies in the server by sending
/lobbies to the server as a command.

Helper functions to convert messages and lobby lists to string
were implemented in this commit as well.