Lobby modal implementation
Lobby modal is correctly positioned in the web view and updates
a table of available lobbies from the server when the user
types /lobbies. Intention is for the lobbies button to
get the lobby list when user clicks on Lobbies.

Need to implement lobby capacity and switch lobbies upon
clicking on one of the table rows, also need to show currently
connected lobby (need to connect user to lobbies in backend).