Fixed welcome message bug
Fixed bug where server wouldn't send welcome message to web
client because it would verify the handshake 'faster' than
the chat client page could render, causing the sent message
to be lost.

Changed it so that the ChatServerHandler will send a welcome
message on the first message it receives from client.

Issue now is to solve the controller double-load problem,
as the initial message is being sent twice by the client.