Lobbies updates for browser client: capacities, formatting
1. Browser client now has the lobby name displayed on top of the list of users
currently connected in the lobby, formatted as a header. To implement this change
ChatServerHandler now reserves the first index of user list arrays with the name
of the lobby that the users are connected to.

2. Implemented capacities for the browser client, so now the current users
connected to the lobby and the total lobby capacity are displayed in the
lobby list. This value updates every time the current user changes
lobbies by clicking or by connecting to another lobby and should
change when any user changes lobby so that it is constantly updating.

In the future, the lobby list should only update capacities (and names)
when the modal window for lobbies is open.

A consequence of the capacity implementation is that the server sends
lobby lists in the following format: lobbyName,usersConnected/capacity

Whereas before it would just send lobbyName for each lobby in the server.