Signup page for web client, server recognizes signup Credentials
1. Created a basic signup page for the web client that is all but
identical to the login page at this point, both in appearance and
in implementation, controllers included.

2. Signup page expectedly sets the signup field of the Credentials
to be submitted to TRUE so that the server can identify that the
credentials are being used for registration as opposed to logging in.

3. Updated ServerAuthHandler to appropriately deal with the new
signup field. Created a new boolean field in ServerAuthHandler,
appropriately named signup. signup defaults to false and is set
by the Credentials received once ServerAuthHandler verifies that
the data that it received was actually a valid FlatBuffers
Credential.