Updated Credentials Schema with new field: signup (bool)
To avoid HTTP/CORS/headaches involved with sending the server
crededentials from the client with custom headers to differetiate
between login and signup credentials, I expanded the Schema's
Credentials table to include a new field, bool signup which is
set to false by default.

The server should not be involved in setting this field, only the
client, so the FlatBuffersCodec in the server was not updated to
account for the change.

Additionally, I overloaded the flatbuffers generated
createCredentials function so that it would be able to create
credentials without providing a boolean to set the signup field
with. Expectedly, this new function automatically sets the field
to false.

By overloading the function, I avoided any problems that might have
occurred with Credentials in the FlatBuffers test file.