Implemented server username registration for web client
1. Created the function verifySignup in LoginAuthorizer which
checks for the validity of a provided username/password
combination in terms of formatting and then inserts the
combination into the database if the username was not already
existing in the database.

2. Edited names of functions in verifySignup to more accurately
express their roles. Specifically, changed the names of the
verifyUser functions to verifyLogin instead.

3. Made sure that each function that dealt with a password char
array in LoginAuthorizer sanitizes the array in memory by filling
the char array with '0' before returning.

4. Extended ServerAuthHandler logic to accommodate the signup
process. Sends HTTP status code 422 on unsuccessful registration
and HTTP status code 200 on success.

5. Changed the signup logic so that it parsed the status code
responses and displayed appropriate messages to the user based
on the received status code.

6. Edited ServerAuthHandler logic providing empty control blocks
which will be used for mobile client responses down the line.