Fixed server timeout bug
There was a bug in the pinger implementation which caused the
server to incorrectly time out users that were active.

In ServerPingHandler, we were sending ping responses when the
channel was not writing after a specific amount of time and
timing out the user if the channel did not read a response from
them after a greater amount of time than the ping response timer.

I took for granted that the client should send a response to every
channel write by the server. Specifically, when other users were
logging in and sending messages, the ServerPing would correctly
not receive any write IdleStates (and thus send no pings), but all
the while the client would not be sending anything in response to
these server writes. Eventually, every client who was not sending
messages to the server would be timed out as the server wouldn't
have read anything from them.

To fix this, I had the server send pings in response to either a
read or write idle state. This way, we guarantee that the client
is aware that the server is still alive and vice-versa. If the
server does not receive either after a time that is longer than
the pinger timer, the server times them out.