Refactoring, finished basic RPS class
Refactored the code so that the Games and game-related classes
that were not involved with the server are in their own package,
conveniently called game.

Finished the basic implementation of the Rock Paper Scissors game.
It accepts two players (users) which are currently implemented as
String. I was going to make a Player class but I think that right
now that is not necessary - it's more important to make sure that
this RPS class will work within the context of the server.

ServerGameHandler is on hold until I can get the server side
challenge/response system working.