Schema extension: GameLobby
Expanded the FlatBuffers Schema to include a new type of Data,
GameLobby. This object will be used to instantiate GameLobby
objects on the server. Currently, we intend to have the web client
package its requests to create lobbies into FlatBuffers serialized
GameLobby objects.

This change was mainly to remove the need for Websockets clients
to send POST requests to create GameLobbies after the WS handshake.
Moreover, it's a lot cleaner to parse a GameLobby object than it is
to parse a text string from clients containing requests to create
game lobbies.

I am realizing that there is a naming conflict between
Schema.GameLobby and the GameLobby class in the Netty backend.
Probably going to change one of their names because this is
already getting confusing.