Game Lobby Creation Dialog for Web Client
Finished frontend for game lobby creation dialog for the web client. Users are able to input a game lobby name, game type, game capacity, and optional password. Client successfully receives all data.

Next step is to package the data into a GameLobby FlatBuffers object and send it to the server.

Used col-xs for this dialog to promote better device cross-compatibility - I should do this for all cols existing in the web client as well.