Refactored Game code for web, new modal for Game lobby
Refactored Game code by separating the Controller logic of ChatGameController into three separate controllers: ChatGameController, GameCreateController, and GameLobbiesController.

ChatGameController is only responsible for switching the HTML of the Game modal on the Chat page. GameLobbiesController is responsible for updating the list of game lobbies and later for sending requests to the server to join games. GameCreateController is responsible for getting and cleaning GameCreationRequest objects from the Web client and sending them to the server for handling.

Currently successful submission of a GameCreationRequest results in the modal switching to a new blank page. This page will represent the Game lobby and should update when new users join/leave the Game lobby. It will also allow the GameLobby host (creator) to start the Game when they feel that enough players have joined and to kick players from the Lobby.

Started a GameService controller which should be responsible for managing the state of the user, i.e. whether or not they are in a game/lobby, name of the game/lobby, etc. This information should be used to restrict their UI capabilities, for instance we shouldn't allow them to create lobbies if they are already in a game or already in a lobby.