Web client updates lobby modal when creating lobby after leave
Had bug where web client would display invalid information about the current game lobby when creating a lobby after leaving one. This was because the object that the modal referred to was cleared by the service after the user left the lobby.

To remedy this, GameService broadcasts an event on the rootScope to notify any listeners that a new game (in this case lobby) has been joined. GameLobbyController updates the Game object accordingly by listening for this event.