Fixed GameLobbyController load bug
Bug fixed by changing GameService _currentGameLobby and _lobbyUsers variables. Before, these variables would replace their references when receiving game lobby updates. Now, they keep their references so that controllers like GameLobbyController can keep track of the changes.

Before, controllers accessing the objects of the GameService would lose data because the references were changed in the service. 

Still had to keep in service broadcasting an event to update changes because the controller needs to know to update the view when the scope variables are updated as well as generate new valid capacities for the game.