Moving modals to ui bootsrap(lobbies), WIP
To adequately address the problem of modals not closing on back,
I think it's appropriate and necessary to convert all modals to
uib (Angular) modals as opposed to pure Bootstrap. I've found
solutions online that will allow me to fix the closing modal
problem if the modals are uib modals already.